enable run-id autocomplete on workplan run
