+++ cstar/cli/workplan/run.py
@@ -7,6 +7,7 @@ import typer
 
 from cstar.base.log import get_logger
 from cstar.cli.workplan.check import check
+from cstar.cli.workplan.shared import list_runs
 from cstar.execution.file_system import is_remote_resource, local_copy
 from cstar.orchestration.dag_runner import build_and_run_dag
 from cstar.orchestration.tracking import TrackingRepository, WorkplanRun
@@ -19,7 +20,10 @@ log = get_logger(__name__)
 def run(
     run_id: t.Annotated[
         str,
-        typer.Option(help="The unique identifier for an execution of the workplan."),
+        typer.Option(
+            help="The unique identifier for an execution of the workplan.",
+            autocompletion=list_runs,
+        ),
     ],
     path: t.Annotated[str, typer.Argument(help="Path to a workplan file.")] = "",
     output_dir: t.Annotated[
