+++ cstar/cli/workplan/status.py
@@ -9,6 +9,7 @@ from rich.panel import Panel
 from rich.table import Table
 
 from cstar.base.log import get_logger
+from cstar.cli.workplan.shared import list_runs
 from cstar.orchestration.tracking import TrackingRepository
 
 log = get_logger(__name__)
@@ -40,29 +41,6 @@ def display_summary(
         table.add_row(open_item, closed_item)
 
     console.print(table)
-
-
-def list_runs(incomplete: str) -> list[tuple[str, str]]:
-    """Retrieve a list of all recorded run-ids.
-
-    Parameters
-    ----------
-    incomplete : str
-        Any value from the user is provided to autocompletion.
-
-    Returns
-    -------
-    t.Iterable[str]
-    """
-    repo = TrackingRepository()
-    run_list = asyncio.run(repo.list_latest_runs(incomplete))
-
-    if not run_list and incomplete:
-        return [(incomplete, "no results found")]
-    elif not run_list:
-        return [("run-id", "no results found")]
-
-    return [(r.run_id, f"Workplan path: {r.workplan_path}") for r in run_list if r]
 
 
 @app.command()
